Move session helpers to backend
